n137/src/main/java/com/snack/leetcode/n137/Solution.java
package com.snack.leetcode.n137;

import java.util.HashMap;
import java.util.Map;

public class Solution {
    public int singleNumber(int[] nums) {
        Map<Integer, Integer> numsmap = new HashMap<>();
        for (int num : nums) {
            numsmap.put(num, numsmap.getOrDefault(num, 0) + 1);
        }
        int ans = 0;
        for (Map.Entry<Integer, Integer> entry : numsmap.entrySet()) {
            if (entry.getValue() == 1) {
                ans = entry.getKey();
                break;
            }
        }
        return ans;
    }
}
